+++ ui/app/src/components/PreviewBackButton/index.ts
@@ -14,6 +14,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { default } from './ForwardButtonStack';
+export { default } from './PreviewBackButton';
 
-export * from './ForwardButtonStack';
+export * from './PreviewBackButton';
