+++ ui/app/src/models/GlobalState.ts
@@ -153,8 +153,8 @@ export interface GlobalState {
     toolsPanelWidth: number;
     pageBuilderPanelWidth: number;
     pageBuilderPanelStack: WidgetDescriptor[];
-    backStack: string[];
-    forwardStack: string[];
+    historyBackStack: string[];
+    historyForwardStack: string[];
     hostSize: WidthAndHeight;
     guest: GuestData;
     assets: PagedEntityState<MediaItem>;
