+++ ui/app/src/state/reducers/preview.ts
@@ -161,8 +161,8 @@ const reducer = createReducer<GlobalState['preview']>(
     toolsPanelWidth: 240,
     pageBuilderPanelWidth: 240,
     pageBuilderPanelStack: [],
-    backStack: [],
-    forwardStack: [],
+    historyBackStack: [],
+    historyForwardStack: [],
     guest: null,
     assets: assetsPanelInitialState,
     audiencesPanel: audiencesPanelInitialState,
